+++ migrations/mysql/005_board_manager.sql
@@ -1,8 +1,8 @@
 -- migrations/mysql/005_board_manager.sql
 -- +goose Up
 CREATE TABLE board_managers (
-    board_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    board_id INT NOT NULL,
+    user_id INT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     
     PRIMARY KEY (board_id, user_id),
